news247_news_capstone_project/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
pymysql>=1.1,<2.0
